update code formatting guidelines

+++ Assets/Project/Source/Math/QuaternionExtensions.cs
@@ -1,14 +1,14 @@
 ﻿using UnityEngine;
 
-namespace Exa.Math
-{
-    public static class QuaternionExtensions
-    {
+namespace Exa.Math {
+    public static class QuaternionExtensions {
         public static Quaternion Multiply(this Quaternion quaternion, float scalar) {
-            return new Quaternion((float) ((double) quaternion.x * (double) scalar),
-                (float) ((double) quaternion.y * (double) scalar),
-                (float) ((double) quaternion.z * (double) scalar),
-                (float) ((double) quaternion.w * (double) scalar));
+            return new Quaternion(
+                (float) (quaternion.x * (double) scalar),
+                (float) (quaternion.y * (double) scalar),
+                (float) (quaternion.z * (double) scalar),
+                (float) (quaternion.w * (double) scalar)
+            );
         }
 
         public static Quaternion RequiredRotation(Quaternion from, Quaternion to) {
@@ -27,10 +27,12 @@ namespace Exa.Math
         }
 
         public static Quaternion Subtract(this Quaternion lhs, Quaternion rhs) {
-            return new Quaternion((float) ((double) lhs.x - (double) rhs.x),
-                (float) ((double) lhs.y - (double) rhs.y),
-                (float) ((double) lhs.z - (double) rhs.z),
-                (float) ((double) lhs.w - (double) rhs.w));
+            return new Quaternion(
+                (float) (lhs.x - (double) rhs.x),
+                (float) (lhs.y - (double) rhs.y),
+                (float) (lhs.z - (double) rhs.z),
+                (float) (lhs.w - (double) rhs.w)
+            );
         }
     }
 }
